requirements.txt
fastapi
uvicorn
pydantic
xgboost
scikit-learn
pandas
numpy
joblib
google-cloud-storage

